feat(SPV-730): refactor files structure
